Input: ALPS - Add code to support "Rushmore" touchpads
These touchpads mostly speak the V3 protocol, but they use a different
method for detection.  The manufacturer's driver follows a sequence that
looks something like:

  Send E7 query and look for an exact match: 73 03 50 ("Dolphin")
  If no match, send EC query and look for 88 {07,08} xx ("Pinnacle")
  If no match, send E7 query and look for 73 02 {0a,14,3c,50} ("Mercury II")
  If no match, send a more complicated query and look for 88 05 xx ("Mercury")

This patch adds the new EC query before the usual E7 query.  It is
presumed (hoped) that anything that isn't in the Pinnacle series will not
return "88 07 xx" or "88 08 xx".

Tested on:

Dell E6230 ("Rushmore" model from the "Pinnacle" series)
Dell E6430 (same as above, but with trackstick + more buttons)

On both PCs, the E7 report is 73 03 0a, and the EC report is 88 08 1d.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>